Compatibility updates for Solr 9.8.1 (#55)
The `IndexSearcher.search(query, collector)` method is now deprecated.
We used custom collectors in several places: sometimes just to count
numbers of results; other times because we wanted to walk a full
result set but didn't care about relevance ranking.

The places using TotalHitCountCollector can be replaced by a call to
IndexSearcher.count(), so no problem there.

For the BibDB, which used to use a custom Collector, I've switched it
to use a standard call to IndexSearcher(query, limit) and to process
the resulting TopDocs.  Using the higher-level API should hopefully be
more future-proof.

It does mean we're now firing the query twice (once to get the result
count, once to accumulate the results into the right-sized array), and
it also means we're now computing relevance ranking scores we don't
use.  But the queries involved are so simple that neither of these
overheads will be noticeable to end users.